adding environment vars renaming components arranging project
