+++ .env.development
@@ -1,0 +1,3 @@
+#APP=customer
+#NAME=whereisitclientapp
+REACT_APP_MAPBOX_TOKEN='pk.eyJ1IjoibmVsc2l0b2dtMjUiLCJhIjoiY2tseHVuenV1MDF3aTJwcnZ3bmVhaXdrNyJ9.xH0-fvi3akdj63xndBuIpg'
