+++ .gitignore
@@ -15,6 +15,7 @@
 .DS_Store
 .env.local
 .env.development.local
+.env.development
 .env.test.local
 .env.production.local
